Use Vim appimage for travis tests
Special handling for Vim AppImage in test_vimrc:
If Vim (not Neovim!) is started and a AppImage is detected
some internal variables are changed to replace the relative path
with the AppImages $APPDIR:
- $VIM
- runtimepath
- helpfile
- packpath

To get the version output durin travis build run
   .../vim.appimage --version 2>/dev/tty

Vim checks that STDERR is connected to a terminal. If that is not the
case, no version output is created.