Fixed handling of comma inside brackets inside braces
'{alpha,[a,]beta}' should become '%(alpha\|\[a\|\]beta\)' as described
in the bash man page.
So if we scan brackets, that are inside braces, the comma should be
handled as a separator for the braces. To prevent this escape the
comma with a backslash:
'{alpha,[a\,]beta}' becomes '%(alpha\|[a,]beta\)'.